+++ lib/tiles/creator/index.js
@@ -1,14 +1,14 @@
-var fileType = require('file-type')
-  , proj4 = require('proj4')
-  , async = require('async')
-  , util = require('util');
+var proj4 = require('proj4')
+  , async = require('async');
 
 var TileBoundingBoxUtils = require('../tileBoundingBoxUtils');
 
 module.exports.initialize = function(width, height, tileMatrix, tileMatrixSet, tileBoundingBox, projectionFrom, projectionTo, canvas, callback) {
   if (typeof(process) !== 'undefined' && process.version) {
+    var LwipTileCreator = require('./lwip');
     new LwipTileCreator(width, height, tileMatrix, tileMatrixSet, tileBoundingBox, projectionFrom, projectionTo, callback);
   } else {
+    var CanvasTileCreator = require('./canvas');
     new CanvasTileCreator(width, height, tileMatrix, tileMatrixSet, tileBoundingBox, projectionFrom, projectionTo, canvas, callback);
   }
 }
@@ -30,27 +30,7 @@ function TileCreator(width, height, tileMatrix, tileMatrixSet, tileBoundingBox, 
   this.sameProjection = projectionFrom.oProj.title === projectionTo.oProj.title;
 }
 
-function CanvasTileCreator(width, height, tileMatrix, tileMatrixSet, tileBoundingBox, projectionFrom, projectionTo, canvas, callback) {
-  TileCreator.apply(this, arguments);
-
-  this.canvas = canvas || document.createElement('canvas');
-  this.canvas.width  = width;
-  this.canvas.height = height;
-  this.ctx = this.canvas.getContext('2d');
-
-  this.image = document.createElement('img');
-
-  this.tileCanvas = document.createElement('canvas');
-  this.tileContext = this.tileCanvas.getContext('2d');
-  this.tileCanvas.width = tileMatrix.tileWidth;
-  this.tileCanvas.height = tileMatrix.tileHeight;
-
-  this.imageData = new Uint8ClampedArray(width * height * 4);
-
-  callback(null, this);
-}
-
-util.inherits(CanvasTileCreator, TileCreator);
+module.exports.TileCreator = TileCreator;
 
 TileCreator.prototype.projectTile = function(tileData, gridColumn, gridRow, callback) {
   var bb = TileBoundingBoxUtils.getTileBoundingBox(this.tileMatrixSet.getBoundingBox(), this.tileMatrix, gridColumn, gridRow);
@@ -121,112 +101,4 @@ TileCreator.prototype.reproject = function (tilePieceBoundingBox, callback) {
       callback();
     }
   );
-};
-
-CanvasTileCreator.prototype.addChunk = function (chunk, xOffset, yOffset) {
-  this.chunks.push({
-    chunk: chunk,
-    x: xOffset,
-    y: yOffset
-  });
-};
-
-CanvasTileCreator.prototype.addPixel = function (targetX, targetY, sourceX, sourceY) {
-  var color = this.tileContext.getImageData(sourceX, sourceY, 1, 1);
-  this.imageData.set(color.data, (targetY * this.width * 4) + (targetX * 4));
-};
-
-CanvasTileCreator.prototype.addTile = function (tileData, gridColumn, gridRow, callback) {
-  var bb = TileBoundingBoxUtils.getTileBoundingBox(this.tileMatrixSet.getBoundingBox(), this.tileMatrix, gridColumn, gridRow);
-
-  var type = fileType(tileData);
-  var binary = '';
-  var bytes = tileData;
-  var len = bytes.byteLength;
-  for (var i = 0; i < len; i++) {
-    binary += String.fromCharCode( bytes[ i ] );
-  }
-  var base64Data = btoa( binary );
-  this.image.onload = function() {
-    this.tileContext.drawImage(this.image, 0, 0);
-    this.projectTile(tileData, gridColumn, gridRow, function() {
-      this.ctx.putImageData(new ImageData(this.imageData, this.width, this.height), 0, 0);
-      async.eachSeries(this.chunks, function(chunk, chunkDone) {
-        var type = fileType(tileData);
-
-        var base64Data = btoa(String.fromCharCode.apply(null, chunk.chunk));
-        var image = document.createElement('img');
-        image.onload = function() {
-          this.ctx.drawImage(image, chunk.x, chunk.y);
-          chunkDone();
-        }.bind(this);
-        image.src = 'data:'+type.mime+';base64,' + base64Data;
-      }.bind(this), callback);
-    }.bind(this));
-  }.bind(this);
-  this.image.src = 'data:'+type.mime+';base64,' + base64Data;
-};
-
-CanvasTileCreator.prototype.getCompleteTile = function (format, callback) {
-  callback(null, this.canvas.toDataURL());
-};
-
-function LwipTileCreator(width, height, tileMatrix, tileMatrixSet, tileBoundingBox, projectionFrom, projectionTo, callback) {
-  TileCreator.apply(this, arguments);
-
-  this.lwip = require('lwip');
-  this.pixels = [];
-  this.lwip.create(this.width, this.height, function(err, image){
-    this.image = image;
-    callback(null, this);
-  }.bind(this));
-}
-
-util.inherits(LwipTileCreator, TileCreator);
-
-LwipTileCreator.prototype.addPixel = function (targetX, targetY, sourceX, sourceY) {
-  this.pixels.push({
-    x: targetX,
-    y: targetY,
-    color: this.tile.getPixel(sourceX, sourceY)
-  });
-};
-
-LwipTileCreator.prototype.addChunk = function (chunk, xOffset, yOffset) {
-  this.chunks.push({
-    chunk: chunk,
-    x: xOffset,
-    y: yOffset
-  });
-};
-
-LwipTileCreator.prototype.addTile = function (tileData, gridColumn, gridRow, callback) {
-  var type = fileType(tileData);
-  this.lwip.open(tileData, type.ext, function(err, tile) {
-    this.tile = tile;
-    this.projectTile(tileData, gridColumn, gridRow, function() {
-      async.eachSeries(this.pixels, function(pixel, pixelDone) {
-        this.image.setPixel(pixel.x, pixel.y, pixel.color, pixelDone);
-      }.bind(this), function(err) {
-        async.eachSeries(this.chunks, function(chunk, chunkDone) {
-          var type = fileType(chunk.chunk);
-          this.lwip.open(chunk.chunk, type.ext, function(err, tile) {
-            tile.crop(Math.abs(Math.min(0, chunk.x)), Math.abs(Math.min(0, chunk.y)), Math.min(this.tileMatrix.tileWidth-1, this.tileMatrix.tileWidth - chunk.x), Math.min(this.tileMatrix.tileHeight-1, this.tileMatrix.tileHeight - chunk.y), function(err, tile) {
-              this.image.paste(Math.max(0, chunk.x), Math.max(0, chunk.y), tile, chunkDone);
-            }.bind(this));
-          }.bind(this));
-        }.bind(this), function(err) {
-          callback(err, this.image);
-        });
-      }.bind(this));
-    }.bind(this));
-  }.bind(this));
-};
-
-LwipTileCreator.prototype.getCompleteTile = function (format, callback) {
-  this.image.writeFile('/tmp/lwip.png', function(err) {
-    this.image.batch()
-    .toBuffer(format, callback);
-
-  }.bind(this));
 };
