+++ lib/tiles/retriever/index.js
@@ -1,4 +1,3 @@
-
 var TileMatrixSetDao = require('../matrixset').TileMatrixSetDao
   , TileBoundingBoxUtils = require('../tileBoundingBoxUtils')
   , TileCreator = require('../creator');
